@@ -1,49 +1,45 @@
 <?php
 
-use Illuminate\Support\Facades\Route;
-
-/*
-|--------------------------------------------------------------------------
-| Web Routes
-|--------------------------------------------------------------------------
-|
-| Here is where you can register web routes for your application. These
-| routes are loaded by the RouteServiceProvider within a group which
-| contains the "web" middleware group. Now create something great!
-|
-*/
+use App\Http\Controllers\indexController;
+use App\Http\Controllers\LoginController;
+use App\Http\Controllers\AgendamentoController;
+use App\Http\Controllers\BancaController;
+use App\Http\Controllers\DocenteController;
+use App\Http\Controllers\ConfigController;
+use App\Http\Controllers\PdfController;
+use App\Http\Controllers\EmailController;
 
 // rotas para login/logout
-Route::get('/','indexController@index')->name('index');
-Route::get('/anteriores','indexController@anteriores')->name('anteriores');
-Route::get('login','LoginController@redirectToProvider')->name('login');
-Route::get('callback', 'LoginController@handleProviderCallback');
-Route::get('logout','LoginController@logout')->name('logout');
+Route::get('/', [indexController::class, 'index'])->name('index');
+Route::get('/anteriores',[indexController::class, 'anteriores'])->name('anteriores');
+Route::get('login',[LoginController::class, 'redirectToProvider'])->name('login');
+Route::get('callback', [LoginController::class, 'handleProviderCallback']);
+Route::get('logout',[LoginController::class, 'logout'])->name('logout');
 
 // rotas de Agendamento de Defesa
-Route::resource('agendamentos','AgendamentoController');
-Route::resource('docentes','DocenteController');
+Route::resource('agendamentos', AgendamentoController::class);
+Route::resource('docentes',DocenteController::class);
 
 // rotas de Banca das Defesas
-Route::get('/agendamentos/{agendamento}/bancas/create','BancaController@create');
-Route::get('/agendamentos/{agendamento}/bancas/{banca}/edit','BancaController@edit');
-Route::patch('/agendamentos/{agendamento}/bancas/{banca}','BancaController@update');
-Route::post('/agendamentos/{agendamento}/bancas','BancaController@store');
-Route::delete('/agendamentos/{agendamento}/bancas/{banca}','BancaController@destroy');
+Route::get('/agendamentos/{agendamento}/bancas/create',[BancaController::class, 'create']);
+Route::get('/agendamentos/{agendamento}/bancas/{banca}/edit',[BancaController::class, 'edit']);
+Route::patch('/agendamentos/{agendamento}/bancas/{banca}',[BancaController::class, 'update']);
+Route::post('/agendamentos/{agendamento}/bancas',[BancaController::class, 'store']);
+Route::delete('/agendamentos/{agendamento}/bancas/{banca}',[BancaController::class, 'destroy']);
 
 // rotas para pdfs
-Route::get('/agendamentos/{agendamento}/{tipo}','PdfController@documentosGerais');
-Route::get('/agendamentos/{agendamento}/bancas/{banca}/{tipo}','PdfController@documentosIndividuais');
-Route::post('/agendamentos/{agendamento}/bancas/{banca}/recibos/proex','PdfController@proex');
-Route::post('/agendamentos/{agendamento}/bancas/{banca}/recibos/proap','PdfController@proap');
-Route::post('/agendamentos/{agendamento}/bancas/{banca}/recibos/passagem','PdfController@passagem');
-Route::post('/agendamentos/{agendamento}/bancas/{banca}/recibos/passagemAuxilio','PdfController@passagemAuxilio');
+Route::get('/agendamentos/{agendamento}/{tipo}',[PdfController::class, 'documentosGerais']);
+Route::get('/agendamentos/{agendamento}/bancas/{banca}/{tipo}',[PdfController::class, 'documentosIndividuais']);
+Route::post('/agendamentos/{agendamento}/bancas/{banca}/recibos/proex',[PdfController::class, 'proex']);
+Route::post('/agendamentos/{agendamento}/bancas/{banca}/recibos/proap',[PdfController::class, 'proap']);
+Route::post('/agendamentos/{agendamento}/bancas/{banca}/recibos/passagem',[PdfController::class, 'passagem']);
+Route::post('/agendamentos/{agendamento}/bancas/{banca}/recibos/passagemAuxilio',[PdfController::class, 'passagemAuxilio']);
 
 // rotas para configs
-Route::get('/configs','ConfigController@edit');
-Route::post('/configs','ConfigController@store');
+Route::get('/configs',[ConfigController::class, 'edit']);
+Route::post('/configs',[ConfigController::class, 'store']);
 
 // rotas para recibos
-Route::post('/agendamentos/{agendamento}/bancas/{banca}/recibos/reciboExterno','EmailController@reciboExterno');
-Route::post('/agendamentos/{agendamento}/bancas/{banca}/recibos/emailDocente','EmailController@emailDocente');
+Route::post('/agendamentos/{agendamento}/bancas/{banca}/recibos/reciboExterno',[EmailController::class, 'reciboExterno']);
+Route::post('/agendamentos/{agendamento}/bancas/{banca}/recibos/emailDocente',[EmailController::class, 'emailDocente']);
 
