@@ -2,7 +2,7 @@
 
 namespace App\Http\Controllers;
 
-use App\Config;
+use App\Models\Config;
 use Illuminate\Http\Request;
 use App\Http\Requests\ConfigRequest;
 
