@@ -7,7 +7,7 @@ use App\Providers\RouteServiceProvider;
 use Illuminate\Foundation\Auth\AuthenticatesUsers;
 
 use Socialite;
-use App\User;
+use App\Models\User;
 use Auth;
 use Illuminate\Http\Request;
 
