@@ -4,7 +4,7 @@ namespace App\Http\Requests;
 
 use Illuminate\Foundation\Http\FormRequest;
 use Illuminate\Validation\Rule;
-use App\Docente;
+use App\Models\Docente;
 class DocenteRequest extends FormRequest
 {
     /**
