@@ -1,16 +1,21 @@
 <?php
 
-namespace App;
+namespace App\Models;
 
+use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
-use App\Docente;
+use App\Models\Docente;
+use App\Models\Agendamento;
+
 class Banca extends Model
 {
+    use HasFactory;
+
     protected $guarded = ['id'];
 
     public function agendamento()
     {
-        return $this->belongsTo('App\Agendamento');
+        return $this->belongsTo(Agendamento::class);
     }
 
     //Função para devolver valores de select
