@@ -1,7 +1,9 @@
 <?php
 
+namespace Database\Seeders;
+
 use Illuminate\Database\Seeder;
-use App\Config;
+use App\Models\Config;
 
 class ConfigSeeder extends Seeder
 {
