@@ -95,7 +95,7 @@
         <br><br><br>
 
         <p class="recuo justificar" style="line-height: 190%;">
-            {!! App\Config::setConfigDeclaracao($agendamento,$bancas,$professor)->declaracao !!}
+            {!! App\Models\Config::setConfigDeclaracao($agendamento,$bancas,$professor)->declaracao !!}
         </p><br><br>
 
         <table width="16cm" style="border='0'; margin-left:4cm; align-items: center; justify-content: center;">
